requirements.txt
python-uinput
attrs
wmctrl
pulsectl
pulsectl_asyncio
mido
python-rtmidi
